server-python/requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv
motor
passlib[bcrypt]